Hometask 7 - Added checks, fixed names

+++ delivery-service/src/main/java/ru/liga/service/CourierService.java
@@ -1,6 +1,7 @@
 package ru.liga.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import ru.liga.dto.CourierDTO;
 import ru.liga.entities.Courier;
@@ -15,7 +16,7 @@ public class CourierService {
 
     private final CourierRepository courierRepository;
 
-    public List<CourierDTO> dtoList () {
+    public List<CourierDTO> getDtoList() {
         return StreamSupport.stream(courierRepository.findAll().spliterator(), false).map(this::courierToDTO).collect(Collectors.toList());
     }
 
@@ -23,23 +24,35 @@ public class CourierService {
         return courierToDTO(getById(id));
     }
 
-    public Courier getById (Integer id){
-        return courierRepository.findById(id).orElseThrow(() -> new RuntimeException("Courier with id = " + id + " is not exists"));
+    public Courier getById(Integer id) {
+        if (id > 0) {
+            return courierRepository.findById(id).orElseThrow(() -> new RuntimeException("Courier with id = " + id + " is not exists"));
+        } else {
+            throw new RuntimeException("Incorrect id");
+        }
     }
 
-    public void convertAndSave (CourierDTO courierDTO) {
+    public void convertAndSave(CourierDTO courierDTO) {
         courierRepository.save(courierToEntity(courierDTO));
     }
 
-    public void deleteById (Integer id) {
-        courierRepository.deleteById(id);
+    public void deleteById(Integer id) {
+        if (id > 0) {
+            try {
+                courierRepository.deleteById(id);
+            } catch (EmptyResultDataAccessException e) {
+                throw new RuntimeException("Courier with id = " + id + " is not exists");
+            }
+        } else {
+            throw new RuntimeException("Incorrect id");
+        }
     }
 
-    public CourierDTO courierToDTO (Courier courier) {
+    public CourierDTO courierToDTO(Courier courier) {
         return new CourierDTO(courier.getId(), courier.getPhone(), courier.getStatus(), courier.getCoordinates());
     }
 
-    public Courier courierToEntity (CourierDTO courierDTO) {
+    public Courier courierToEntity(CourierDTO courierDTO) {
         return new Courier(courierDTO.getPhone(), courierDTO.getStatus(), courierDTO.getCoordinates());
     }
 }
