+++ order-service/src/main/java/ru/liga/service/OrderService.java
@@ -1,6 +1,7 @@
 package ru.liga.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import ru.liga.dto.OrderDTO;
 import ru.liga.entities.Order;
@@ -17,7 +18,7 @@ public class OrderService {
 
     private final OrderRepository orderRepository;
 
-    public List<OrderDTO> dtoList() {
+    public List<OrderDTO> getDtoList() {
         return StreamSupport.stream(orderRepository.findAll().spliterator(), false).map(this::orderToDTO).collect(Collectors.toList());
     }
 
@@ -26,31 +27,47 @@ public class OrderService {
     }
 
     public Order getById(Integer id) {
-        return orderRepository.findById(id).orElseThrow(() -> new RuntimeException("Order with id = " + id + " is not exists"));
+        if (id > 0) {
+            return orderRepository.findById(id).orElseThrow(() -> new RuntimeException("Order with id = " + id + " is not exists"));
+        } else {
+            throw new RuntimeException("Incorrect id");
+        }
     }
 
-    public void convertAndSave(OrderDTO orderDTO){
+    public void convertAndSave(OrderDTO orderDTO) {
         orderRepository.save(orderToEntity(orderDTO));
     }
 
     public void deleteById(Integer id) {
-        orderRepository.deleteById(id);
-    }
-
-    public void updateStatusById(Integer id, OrderStatus status) {
-        Optional<Order> row = orderRepository.findById(id);
-        if (row.isPresent()) {
-            Order order = row.get();
-            order.setStatus(status);
-            orderRepository.save(order);
+        if (id > 0) {
+            try {
+                orderRepository.deleteById(id);
+            } catch (EmptyResultDataAccessException e) {
+                throw new RuntimeException("Order with id = " + id + " is not exists");
+            }
+        } else {
+            throw new RuntimeException("Incorrect id");
         }
     }
 
-    public OrderDTO orderToDTO (Order order) {
+    public void updateStatusById(Integer id, OrderStatus status) {
+        if (id > 0) {
+            Optional<Order> row = orderRepository.findById(id);
+            if (row.isPresent()) {
+                Order order = row.get();
+                order.setStatus(status);
+                orderRepository.save(order);
+            }
+        } else {
+            throw new RuntimeException("Incorrect id");
+        }
+    }
+
+    public OrderDTO orderToDTO(Order order) {
         return new OrderDTO(order.getId(), order.getCustomerId(), order.getRestaurantId(), order.getStatus(), order.getCourierId(), order.getTimestamp());
     }
 
-    public Order orderToEntity (OrderDTO orderDTO) {
+    public Order orderToEntity(OrderDTO orderDTO) {
         return new Order(orderDTO.getCustomerId(), orderDTO.getRestaurantId(), orderDTO.getStatus(), orderDTO.getCourierId(), orderDTO.getTimestamp());
     }
 }
